Move Connection and ConnectionFactory implementations into the impl package
