+++ src/main/java/com/suse/saltstack/netapi/client/impl/RequestFactory.java
@@ -1,4 +1,4 @@
-package com.suse.saltstack.netapi.client;
+package com.suse.saltstack.netapi.client.impl;
 
 import com.suse.saltstack.netapi.config.ClientConfig;
 import static com.suse.saltstack.netapi.config.ClientConfig.*;
